add ControlMessageHandler
